service-workers/service-worker/skip-waiting-using-registration.https.html

<!DOCTYPE html>
<title>Service Worker: Skip waiting using registration</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
'use strict';

promise_test(async t => {
  const SCOPE = 'resources/blank.html?skip-waiting-using-registration';
  const SCRIPT1 = 'resources/empty.js';
  const SCRIPT2 = 'resources/skip-waiting-worker.js';
  
  let registration =
      await service_worker_unregister_and_register(t, SCRIPT1, SCOPE);
  await wait_for_state(t, registration.installing, 'activated');
  const frame = await with_iframe(SCOPE);
  const new_serviceWorker = await frame.contentWindow.navigator.serviceWorker;
  assert_equals(new_serviceWorker.controller.scriptURL, normalizeURL(SCRIPT1),
      'Document controller scriptURL should equal to the first one');
  registration = await navigator.serviceWorker.register(SCRIPT2,
                                                        {scope:SCOPE});
  const event = await new Promise(resolve => {
    new_serviceWorker.oncontrollerchange = resolve;
  });
  assert_equals(event.type, 'controllerchange', 'controllerchange');
  assert_true(
      event.target instanceof frame.contentWindow.ServiceWorkerContainer,
      'ServiceWorkerContainer');
  assert_equals(event.target.controller.state, 'activating',
                'Controller state should be activating');
  assert_equals(new_serviceWorker.controller.scriptURL, normalizeURL(SCRIPT2),
      'Controller scriptURL should change to the second one');

  t.add_cleanup(async() => {
    if (registration)
      await registration.unregister();
    if (frame)
      frame.remove();
  });

  assert_not_equals(registration.active, null,
                    'Registration active worker should not be null');
  fetch_tests_from_worker(registration.active);

}, 'Test skipWaiting while a client is using the registration');

</script>
